[issue512] remove TODO

+++ src/search/task_proxy.h
@@ -296,8 +296,7 @@ public:
     using ItemType = FactProxy;
     explicit State(const AbstractTask &task_, std::vector<int> &&values_)
         : task(task_), values(values_) {
-        // TODO: Let AbstractTask::get_num_* methods return size_t.
-        assert(size() == static_cast<std::size_t>(task.get_num_variables()));
+        assert(static_cast<int>(size()) == task.get_num_variables());
     }
     ~State() {}
     std::size_t size() const {
